Refactor to use RTOS tasks
Significant update. Moved the vlx-sampler to its own task, and the eink updater to another task.

There is still a todo to encapsulate the lidar value, and array of values into a mutex-controlled class.

The eInk display will now show an error msg if the i2c-vlx fails to init.